legacy_python/requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
passlib[bcrypt]
itsdangerous
google-generativeai
python-dotenv
